feat(mcp-analytics): use shared LemonUI date filter, add flag-aware entrypoint
Make `DateFilter` a flag-aware entrypoint that renders the existing LemonUI picker (extracted to `LemonDatePicker`) by default, or a Quill picker when the `DATEPICKER_COMPONENT` multivariate flag is `quill`. Switch MCP analytics off its bespoke Quill date filter onto the shared `DateFilter` so it matches the rest of the app. Document the Quill `DateTimePicker` as incomplete in the agent guides.

Generated-By: PostHog Code
Task-Id: b3067154-0b98-437d-aff8-cb4105495379
