+++ products/mcp_analytics/frontend/MCPAnalyticsToolQuality.tsx
@@ -6,6 +6,7 @@ import { type ChartTheme } from '@posthog/quill-charts'
 import { Button } from '@posthog/quill-primitives'
 
 import { buildTheme } from 'lib/charts/utils/theme'
+import { DateFilter } from 'lib/components/DateFilter/DateFilter'
 import { TagsCombobox } from 'lib/components/Scenes/TagsCombobox'
 import { LinkPrimitive } from 'lib/lemon-ui/Link/Link'
 import { Tooltip } from 'lib/lemon-ui/Tooltip'
@@ -14,7 +15,6 @@ import { urls } from 'scenes/urls'
 
 import { themeLogic } from '~/layout/navigation-3000/themeLogic'
 
-import { McpDateFilter } from './components/McpDateFilter'
 import { mcpAnalyticsToolQualityLogic } from './mcpAnalyticsToolQualityLogic'
 import { ToolQualityCharts } from './tool-quality/ToolQualityCharts'
 import { ToolQualityTable } from './tool-quality/ToolQualityTable'
@@ -38,11 +38,10 @@ function FilterBar(): JSX.Element {
                     dataAttr="mcp-tool-quality-category-scope"
                 />
             </div>
-            <McpDateFilter
+            <DateFilter
                 dateFrom={dateFilter.dateFrom}
                 dateTo={dateFilter.dateTo}
                 onChange={(dateFrom, dateTo) => setDateFilter(dateFrom, dateTo)}
-                dataAttr="mcp-tool-quality-date-filter"
             />
             {hasScope && sharePct !== null ? (
                 <Tooltip
